requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
openpyxl==3.1.2
python-docx==0.8.11
Jinja2==3.1.2
python-multipart==0.0.6
